+++ helm-app-operator/pkg/helm/engine/ownerref.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package helm
+package engine
 
 import (
 	"fmt"
